[REF] web: move fail-to-fetch error handler to core
Before this commit, the fail-to-fetch error handler was located in the
`webclient` folder. This handler logically belongs in the `offline`
folder within `core`.

This commit moves the file to its proper location.
